Fixing Recreations:
Creating Class Recreations to hold the recreation Items for the Cinema with autoprice levesl.

Now the Cinema, the Shopping Mall and the Fast Food are of the same class.
Now the Ninja have a single method to take any kind of items and to use it as of its purpose.

As a result we have higher level of abstraction and polymorphism.